+++ controlplane/src/core/repositories/SubgraphRepository.ts
@@ -807,17 +807,7 @@ export class SubgraphRepository {
 
     const checkList = await this.db.query.schemaChecks.findMany({
       columns: {
-        id: true,
-        targetId: true,
-        createdAt: true,
-        isComposable: true,
-        isDeleted: true,
-        hasBreakingChanges: true,
-        hasClientTraffic: true,
-        forcedSuccess: true,
-        ghDetails: true,
-        hasLintErrors: true,
-        hasGraphPruningErrors: true,
+        proposedSubgraphSchemaSDL: false,
       },
       limit,
       offset,
@@ -855,6 +845,9 @@ export class SubgraphRepository {
           : undefined,
         hasLintErrors: c.hasLintErrors ?? false,
         hasGraphPruningErrors: c.hasGraphPruningErrors ?? false,
+        clientTrafficCheckSkipped: c.clientTrafficCheckSkipped ?? false,
+        lintSkipped: c.lintSkipped ?? false,
+        graphPruningSkipped: c.graphPruningSkipped ?? false,
       })),
       checksCount,
     };
@@ -954,6 +947,9 @@ export class SubgraphRepository {
         : undefined,
       hasLintErrors: check.hasLintErrors ?? false,
       hasGraphPruningErrors: check.hasGraphPruningErrors ?? false,
+      clientTrafficCheckSkipped: check.clientTrafficCheckSkipped ?? false,
+      lintSkipped: check.lintSkipped ?? false,
+      graphPruningSkipped: check.graphPruningSkipped ?? false,
     };
   }