+++ router/pkg/config/config.go
@@ -2,14 +2,14 @@ package config
 
 import (
 	"fmt"
-	"github.com/wundergraph/cosmo/router/internal/unique"
 	"os"
 	"time"
 
-	"github.com/goccy/go-yaml"
-
 	"github.com/caarlos0/env/v11"
+	"github.com/goccy/go-yaml"
 	"github.com/joho/godotenv"
+
+	"github.com/wundergraph/cosmo/router/internal/unique"
 	"github.com/wundergraph/cosmo/router/pkg/otel/otelconfig"
 )
 
@@ -74,6 +74,7 @@ type PropagationConfig struct {
 	Jaeger       bool `yaml:"jaeger"`
 	B3           bool `yaml:"b3"`
 	Baggage      bool `yaml:"baggage"`
+	Datadog      bool `yaml:"datadog"`
 }
 
 type Prometheus struct {
@@ -161,6 +162,17 @@ type BackoffJitterRetry struct {
 	MaxAttempts int           `yaml:"max_attempts" envDefault:"5"`
 	MaxDuration time.Duration `yaml:"max_duration" envDefault:"10s"`
 	Interval    time.Duration `yaml:"interval" envDefault:"3s"`
+}
+
+type SubgraphCacheControlRule struct {
+	Name  string `yaml:"name"`
+	Value string `yaml:"value"`
+}
+
+type CacheControlPolicy struct {
+	Enabled   bool                       `yaml:"enabled" envDefault:"false" env:"CACHE_CONTROL_POLICY_ENABLED"`
+	Value     string                     `yaml:"value" env:"CACHE_CONTROL_POLICY_VALUE"`
+	Subgraphs []SubgraphCacheControlRule `yaml:"subgraphs,omitempty"`
 }
 
 type HeaderRules struct {
@@ -609,17 +621,27 @@ type AccessLogsFileOutputConfig struct {
 	Path    string `yaml:"path" env:"ACCESS_LOGS_FILE_OUTPUT_PATH" envDefault:"access.log"`
 }
 
+type ApolloCompatibilityFlags struct {
+	EnableAll       bool                               `yaml:"enable_all" envDefault:"false" env:"APOLLO_COMPATIBILITY_ENABLE_ALL"`
+	ValueCompletion ApolloCompatibilityValueCompletion `yaml:"value_completion"`
+}
+
+type ApolloCompatibilityValueCompletion struct {
+	Enabled bool `yaml:"enabled" envDefault:"false" env:"APOLLO_COMPATIBILITY_VALUE_COMPLETION_ENABLED"`
+}
+
 type Config struct {
 	Version string `yaml:"version,omitempty" ignored:"true"`
 
-	InstanceID     string           `yaml:"instance_id,omitempty" env:"INSTANCE_ID"`
-	Graph          Graph            `yaml:"graph,omitempty"`
-	Telemetry      Telemetry        `yaml:"telemetry,omitempty"`
-	GraphqlMetrics GraphqlMetrics   `yaml:"graphql_metrics,omitempty"`
-	CORS           CORS             `yaml:"cors,omitempty"`
-	Cluster        Cluster          `yaml:"cluster,omitempty"`
-	Compliance     ComplianceConfig `yaml:"compliance,omitempty"`
-	TLS            TLSConfiguration `yaml:"tls,omitempty"`
+	InstanceID     string             `yaml:"instance_id,omitempty" env:"INSTANCE_ID"`
+	Graph          Graph              `yaml:"graph,omitempty"`
+	Telemetry      Telemetry          `yaml:"telemetry,omitempty"`
+	GraphqlMetrics GraphqlMetrics     `yaml:"graphql_metrics,omitempty"`
+	CORS           CORS               `yaml:"cors,omitempty"`
+	Cluster        Cluster            `yaml:"cluster,omitempty"`
+	Compliance     ComplianceConfig   `yaml:"compliance,omitempty"`
+	TLS            TLSConfiguration   `yaml:"tls,omitempty"`
+	CacheControl   CacheControlPolicy `yaml:"cache_control_policy"`
 
 	Modules        map[string]interface{} `yaml:"modules,omitempty"`
 	Headers        HeaderRules            `yaml:"headers,omitempty"`
@@ -668,6 +690,7 @@ type Config struct {
 	StorageProviders          StorageProviders          `yaml:"storage_providers"`
 	ExecutionConfig           ExecutionConfig           `yaml:"execution_config"`
 	PersistedOperationsConfig PersistedOperationsConfig `yaml:"persisted_operations"`
+	ApolloCompatibilityFlags  ApolloCompatibilityFlags  `yaml:"apollo_compatibility_flags"`
 }
 
 type LoadResult struct {