+++ router/CHANGELOG.md
@@ -4,6 +4,28 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.115.1](https://github.com/wundergraph/cosmo/compare/router@0.115.0...router@0.115.1) (2024-09-25)
+
+### Bug Fixes
+
+* handle empty query in plan ([#1219](https://github.com/wundergraph/cosmo/issues/1219)) ([37af012](https://github.com/wundergraph/cosmo/commit/37af0123d8185897a5842616377be320b696037c)) (@thisisnithin)
+
+# [0.115.0](https://github.com/wundergraph/cosmo/compare/router@0.114.1...router@0.115.0) (2024-09-24)
+
+### Features
+
+* enable datadog trace propagation ([#1204](https://github.com/wundergraph/cosmo/issues/1204)) ([1747bf5](https://github.com/wundergraph/cosmo/commit/1747bf53e5ccbd92e323cead8dfb2adccbfe3b7b)) (@df-wg)
+
+## [0.114.1](https://github.com/wundergraph/cosmo/compare/router@0.114.0...router@0.114.1) (2024-09-23)
+
+**Note:** Version bump only for package router
+
+# [0.114.0](https://github.com/wundergraph/cosmo/compare/router@0.113.0...router@0.114.0) (2024-09-23)
+
+### Features
+
+* add apollo compatibility mode with support for valueCompletion ([#1205](https://github.com/wundergraph/cosmo/issues/1205)) ([18b1ef0](https://github.com/wundergraph/cosmo/commit/18b1ef01b12945d2f3acc80ea9548a17f9effa21)) (@jensneuse)
+
 # [0.113.0](https://github.com/wundergraph/cosmo/compare/router@0.112.0...router@0.113.0) (2024-09-20)
 
 ### Bug Fixes